Small commit Just a small commit I thought i need to make so I don't lose any changes if something back happend.

+++ src/assetManager/include/assetDatas/TextureData.h
@@ -11,10 +11,10 @@
 namespace am
 {
     struct TextureData {
-        std::vector<std::uint8_t> pixels;
-        int width{0};
-        int height{0};
-        int channels{0};
+        std::vector<std::uint32_t> pixels;
+        uint32_t width{0};
+        uint32_t height{0};
+        uint32_t channels{0};
         bool hasAlpha{false};
     };
 }
